feat(Tracking): add device details records for steam devices
The new SteamVRInputSourceRecord provides a very basic implementation
of the device details record for the SteamVR CameraRig. It doesn't
currently use any of the SteamVR library functionality to determine
details and instead relies on what is provided by the Unity libraries.

This should be updated in the future as this is just a temporary
solution.